extract player class
